requirements/ci.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    ./bin/compile_dependencies.sh
#
asgiref==3.5.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django
astroid==2.9.3
    # via pylint
attrs==21.4.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   jsonschema
beautifulsoup4==4.10.0
    # via webtest
certifi==2021.10.8
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   elastic-apm
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.12
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   requests
coverage==4.5.4
    # via -r requirements/test-tools.in
cssselect==1.1.0
    # via pyquery
deprecated==1.2.13
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   redis
django==3.2.12
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-axes
    #   django-choices
    #   django-formtools
    #   django-hijack
    #   django-jenkins
    #   django-otp
    #   django-phonenumber-field
    #   django-redis
    #   django-rosetta
    #   django-sniplates
    #   djangorestframework
    #   drf-extra-fields
    #   drf-spectacular
    #   maykin-django-two-factor-auth
django-admin-index==1.5.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-axes==5.31.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-choices==1.7.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-formtools==2.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   maykin-django-two-factor-auth
django-hijack==3.1.6
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-ipware==4.0.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-axes
django-jenkins==0.110.0
    # via -r requirements/test-tools.in
django-ordered-model==3.5
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-admin-index
django-otp==1.1.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   maykin-django-two-factor-auth
django-phonenumber-field==5.2.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   maykin-django-two-factor-auth
django-redis==5.2.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-rosetta==0.9.8
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-sniplates==0.7.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-webtest==1.9.9
    # via -r requirements/test-tools.in
djangorestframework==3.13.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   drf-extra-fields
    #   drf-spectacular
drf-extra-fields==3.2.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
drf-spectacular==0.21.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
elastic-apm==6.7.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
factory-boy==3.2.1
    # via -r requirements/test-tools.in
faker==13.2.0
    # via factory-boy
freezegun==1.1.0
    # via -r requirements/test-tools.in
idna==3.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   requests
importlib-resources==5.4.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   jsonschema
inflection==0.5.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   drf-spectacular
isort==5.10.1
    # via pylint
jsonschema==4.4.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   drf-spectacular
lazy-object-proxy==1.7.1
    # via astroid
lxml==4.8.0
    # via pyquery
maykin-django-two-factor-auth==2.0.4
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
mccabe==0.6.1
    # via pylint
packaging==21.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   redis
pep8==1.7.1
    # via -r requirements/test-tools.in
phonenumbers==8.12.43
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
pillow==9.0.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
platformdirs==2.5.1
    # via pylint
polib==1.1.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-rosetta
psycopg2==2.9.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
pylint==2.12.2
    # via -r requirements/test-tools.in
pyparsing==3.0.7
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   packaging
pyquery==1.4.3
    # via -r requirements/test-tools.in
pyrsistent==0.18.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   jsonschema
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
python-decouple==3.6
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
python-dotenv==0.19.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
pytz==2021.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
pyyaml==6.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   drf-spectacular
qrcode==6.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   maykin-django-two-factor-auth
redis==4.1.4
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-redis
requests==2.27.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-rosetta
    #   requests-mock
requests-mock==1.9.3
    # via -r requirements/test-tools.in
sentry-sdk==1.5.6
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
six==1.16.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-choices
    #   python-dateutil
    #   qrcode
    #   requests-mock
soupsieve==2.3.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django
tblib==1.7.0
    # via -r requirements/test-tools.in
toml==0.10.2
    # via pylint
typing-extensions==4.1.1
    # via
    #   astroid
    #   pylint
uritemplate==4.1.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   drf-spectacular
urllib3==1.26.8
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.20
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
waitress==2.0.0
    # via webtest
webob==1.8.7
    # via webtest
webtest==3.0.0
    # via django-webtest
wrapt==1.13.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   astroid
    #   deprecated
zipp==3.7.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
setuptools==60.9.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   astroid
    #   django-axes
